+++ CHANGES.md
@@ -1,6 +1,7 @@
 # unreleased
 - Support nested `let..in` for `[%sedlex.regexp?]` definitions
 - Add support for named captured group (#177)
+- Add support for `when` guard expressions
 
 # 3.7 (2025-10-06)
 - Update to unicode 17.0.0